bring over RegisterUser from CMS
